Fix course cart bug
The bug seems to stem from the fact that doing a check like if (course.info) is throwing an exception when course is null and aborting functions prematurely.  Fixing the if statements seems to fix the functionality.  I also changed == usage to === and var usage to let, and ran an autoformatter.